tools/codegen/requirements.txt
jupyterlab
pycparser
tqdm>=4.42.0
ipywidgets
jupytext
pyelftools
jinja2
